restructure rust sources under `src` directory
Initially, the sources were splitted in many packages to speed up
compilation (see https://github.com/KnpLabs/should-skip-ci/pull/2 ).
This has lead to difficulties to navigate through the source code, and
created complexity in managing version bumps due to many Cargo.toml
files.

This commit organizes back the structure of the code under the `src`
dir, and bumps rust edition from 2018 to 2021 which would avoid long
re-compilations, which were the cause of the initial split.
