--- cli/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "cli"
-version = "0.2.1"
-authors = ["KNP Labs"]
-edition = "2018"
-
-[dependencies]
-structopt = "0.3"
