--- logger/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "logger"
-version = "0.2.1"
-authors = ["KNP Labs"]
-edition = "2018"
-
-[dependencies]
-log = "0.4"
-simple_logger = { version = "2.1.0", default-features = false }
