--- utils/Cargo.toml
@@ -1,5 +1,0 @@
-[package]
-name = "utils"
-version = "0.2.1"
-authors = ["KNP Labs"]
-edition = "2018"
